improving the script to benchmark the grid sizes [skip ci]
