+++ .gitignore
@@ -217,6 +217,7 @@ benchmarks/debug_speed_improvments.py
 perf.data
 perf.data.old
 test_profile.py.lprof
+benchmarks/case14.json
 benchmarks/case9241pegase.json
 benchmarks/case118.json
 benchmarks/case_illinois200.json
@@ -228,6 +229,22 @@ benchmarks/case2848rte.json
 benchmarks/case2869pegase.json
 benchmarks/case3120sp.json
 benchmarks/case6495rte.json
+benchmarks/case6515rte.json
+benchmarks/case9241pegase.json.old
+benchmarks/case118.json.old
+benchmarks/case_illinois200.json.old
+benchmarks/case300.json.old
+benchmarks/case1354pegase.json.old
+benchmarks/case1888rte.json.old
+benchmarks/GBnetwork.json.old
+benchmarks/case2848rte.json.old
+benchmarks/case2869pegase.json.old
+benchmarks/case3120sp.json.old
+benchmarks/case6495rte.json.old
+benchmarks/case6515rte.json.old
+benchmarks/benchmark.png
+benchmarks/benchmark2.png
+cktso
 _pandapower
 benchmarks/slacks.py
 _gridcal
